Add nerve readiness check
This is a bit beyond an MVP so is behind a feature flag. Technically we
want this if we want to mimic the current bounce logic. Without a
readiness check that asserts a service is UP in smartstack then the
bounce could kill old instances before the new ones can serve traffic.